Main file to parse args
Adds input flags to determine if we should run a client or server.

Since this is the first commit with main.py we're also adding a shebang
to the file, and chmod a+xing it. I'm pretty sure that git stores file
modes, so make sure you commit your chmods!