Commands
This adds a "command" class that can serialize and deserialize. The
reason I didn't just use pickle (or jsonpickle) is that they will happly
load executable code, which we definitely don't want to toss around over
the network on our servers or the machines of end-users. The reason we
don't just serialize a bunch of commands as json in one big go is that I
want to leave it open to implementing different serialization methods if
network bandwidth becomes a pain point.

The only current tradeoff is that the inheritor needs to spell out which
class members are needed to recreate the object. This is less than
ideal, but it does avoid doing crazy inferences several times in the
middle of the game loop.