Builds out client/server communication
Adds a thread for the client so it can recieve asynchronously. Adds a
send queue for the server. The client and server code is looking so
similar that it might make sense to refactor them - maybe we have an
asynch transciever class that has a send queue and a recieve queue.

As a simple demo, we're drawing a circle on-screen on every click
(sending it through the network as a ping command)