Large refactor in Net
Abstracts the multithreaded nature of the client and server into the
messenger. This makes the rest of the code signifigantly easier to read
- the client and the server both use messengers, the client is
  (currently) just a wrapper over messenger and the server uses a single
thread to first accept connections then go into a loop which checks the
queues of each messenger and, once it's recieved from each one on a
given step (using the same logic as the previous incarnation) sends the
step to each client (via the messenger) and deletes the step.