Implements "step ahead" delagging concept
This transmits steps with future step-uids so that we have a buffer of
steps to execute. This is used to smooth out what might be a laggy
connection.

Making it possible to change lag compensation mid-game would be more
tricky - we would need to insert padding steps on every client.