Sprites / data loading
This sets up a data loader that looks for a 'sprites.json' and after
another call loads images. It also adds a sprite class. Note that this
is not a sprite-per-entity class, but belongs to the sprite draw system,
which draws sprites for each entity. I'm not sure that this is the most
optimal approach, but it's what I've got so far.

Note that this commit adds a graphic by DanC of http://lostgarden.com.
It's from the Tyrian collection (A tank that I cropped from a sprite
sheet) and it's under the lost garden liscense (equivalent to CC_BY)