Set based hashing
This creates a set from the items in our dicts rather than dumping to
json so that we get a deterministic hash. It uses python's builtin
'hash' function rather than md5. Because of this, we need to use
hashable types as components for our entities - this is a bit of a pain.
I'll need to look into if you can implement __to_hash or some such to
make a custom type hashable.