selection, filters
The basic skeleton of how selection will work. To implement this, I've
added a notion of 'filters' which are basically queries for the ECS. In
this case, we have one that gets anything within the selection box, but
they could also be used to decide what a system should act on (I have
half a mind to replace the criteria system for Systems with Filters) as
well as stuff like area-of-effect weapons. It could also be possible to
chain filters, for example, to implement
double-click-select-all-of-a-type.