Primative movement
This adds a very primative form of flying movement to the game. Mostly
to show that the system can do something like this - the
'angle_to_frame' code in this commit is less than stellar (it uses
degrees as an intermediate so I could wrap my head around it, leaves
print statements in, etc.)

More dangerously, it uses floating point math. Just as a placeholder,
but still, it may be a risk for out of sync issues - that needs further
testing.