Support YAML
JSON is a very nice format. Supporting it for data assets is a good
idea. However the prospect of actually hand-writing all of the data for
yet another game in json is sort of untenable. Writing YAML is much
easier as a human, so the data loader will now open both json and
yaml/yml files. XML is another possible format to support, but it does
not parse as cleanly into the sort of trees that I'd like to make for
these data files.