Spawn a building at start
This adds a building unit type which spawns at the start location. It
adds a sprite (from Tyrian again, by danc) for it. Also moving towards
the ability to spawn units - there's a 'q' command for the building to
spawn a tank. I think I'm going to go with hotkeys rather than buttons
for the forseeable future because I don't feel like implementing a whole
GUI just yet.