Adds 'make' command
This adds a command for making units, and hooks up the GUI to use it.
There is a good chunk of MVPish code that decides if the units are going
to accept the hotkey - this should probably be moved to a one-time thing
that is called when the selection set changes (so that buttons can be
displayed to the user).

But for now it's logically consistant and works. Press q. Spawn a tank.
Drive it around. Live.