Factions
Adds a notion of factions to the data setup. The only parameter for
factions is what building they spawn at the start location, but it could
also be used to, for example, customize the HUD (note that the hud is
created in the handshake function) soundtrack, etc.

This involves some renaming which ends up making this a pretty big
commit.