Targeting a unit works (after a fashon)
This (somewhat heavy) commit adds a way to dispatch different commands
depending on if the user has clicked a unit or a location. The actual
change is quite subtile - to see any effect, you must use 'm' to get the
crosshairs mouse then click on the outer edge of a sprite - the tank (or
sidecar) will move to the center of that sprite.