Adds Tiled Library and some maps, danc graphics
Now the server chooses what map to use and sends it to the clients in
the handshake. Clients use pytmx to open the maps. Re-uses a map and
tileset made for Scandium (and yes, the name is a Knife Party reference)
though the tileset is, I should stress, from DanC's famous Hard Vaccum
tileset.