Adds hacky scrolling
This adds a less than ideal method of the GUI communicating scrolling to
the game. It's good enough to use for testing, however, so it stays for
now. Now, rather than just returning a command or null, handle event can
return:

Null

Null, new offset

Command, new offset

Any of these are valid, but obviously it's a bit hacky. It also ties
movement to the event handler, which means you need to hammer the arrow
keys to move around.

This also adds a system to draw the move goals of selected units - this
aids in debugging, but might also be a neat gameplay feature.