Units follow paths
If an entity has a 'path' member, it will now be moved by the
PathFollowSystem. It is similar to the fly system, but rather than
moving straight to the goal position, it moves to each position in the
path in turn. It is designed to accomodate units being pushed around by
flocking behavior.