More natural paths
By slicing the first and last node off of the paths, we make pathing
more natural. Since "closest_node" is giving us the node that the unit
either is already in or wants to get to, there's no sense in leaving it
in the path proper. Having it first move to the center of the goal tile
then back out to the exact location was awkward.